+++ docs/release_notes.rst
@@ -1,10 +1,30 @@
 Release Notes
 =============
 
-v4.5.0
+v6.0.0
 ------
 
 * Cache authentication credentials for performance.
+* Integration tests are part of package's tests
+* Addition of `swift.ConflictError` exception, which is thrown when storage
+  nodes have consistency issue deleting container
+* Retry logic on `SwiftPath.rmtree` so that container deletes will retry
+  when hitting a `swift.ConflictError`
+* Swift conditions are now functions that take results and return a Boolean.
+* `SwiftPath.download` returns a list of downloaded paths
+
+v5.0.0
+------
+
+* Default `SwiftPath.temp_url` to using ``inline=True``, since that's the general
+  expectation for how we've used Apache filer / generally you expect URLs to
+  render in-browser rather than default to attachment.
+
+v4.5.0
+------
+
+* Added `SwiftPath.temp_url` for generating temporary object URLs
+* Added environment variable ``OS_TEMP_URL_KEY`` for pulling in default temp url key
 
 v4.4.1
 ------